Fix i18n issue with "Paste the configuration here"
In 552d6c04 this string was added as "Paste Configuration here"
in Base.lproj/Localizable.strings, while in the source code,
ExportImportViewController, the string is
"Paste the configuration here".

This change updates the key in Base.lproj as well as in translations.
After a rebuild, that should put the translations correctly in the
software, I presume. That is, assuming this i18n framework works
similar to gettext, where the original string is the key.